Take 2
Previous fix was not complete as it had problem when feeds there was no channel created in Disguise to match a camera in the scene.  Now we use the name of camera and remote parameter to find the object to display (instead of relying on have exactly the same number of objects exactly in the same order).